fix: use empty value when execution does not exist yet (#65)
Currently, if an input is defined from the last execution of the stage,
and an execution does not exist, we error out. This requires the user to
create/update the stage in a not-so-clear and specific sequence of
steps. A better initial behavior is to just use an empty value if an
execution does not exist yet.