refactor: new integration / executor should not require API / core code changes (#80)
Adding a new integration or executor requires updating the API and some
logic in the workers for the integration/executor to work properly. That
shouldn't be the case. Ideally, adding a new integration/executor should
be just about adding a new implementation for the interface you want,
and registering it. The changes in this pull request allow for that to
happen.

#### Key Changes

- Introduced pkg/registry package to centrally manage integrations and
executors using factory patterns
- Moved integration-specific executors from pkg/executors/ to
pkg/integrations/ for better organization. Executors that don't require
an integration are still placed in pkg/executors/.
- Updated the API to use generic types for the execution spec. That way,
the executor is the one responsible for what the spec looks like, and
when adding a new executor, we don't need to update the API.
- Moved executor spec validation logic into the executor interface
- Moved the integration webhook setup logic into the integration
interface