refactor: relationship between users and organizations (#99)
### Changes

- The relationship between users and organization was updated, with a
new account model being used. Users are now a concrete association
between an account and an organization.
- Currently, we are using an `is_active` flag on the users table to
determine if a user is part of the organization or it it has been only
invited. That is confusing and prone to errors. Now, we use a
`organization_invitations` table to keep track of the invitations, and
once an account that has been invited logs in, we add them to the
organizations that have invited it. This change also makes it easier for
us to extend things in the future, and make invitations expire, or
support invitations based on tokens that can be used by anyone.
- Removed /auth/me, /auth/token/exchange, and
/auth/{provider}/disconnect endpoints. the CLI will be updated to use
user API tokens in a following pull request.
- The API routes were updated to expect an account cookie and an
`x-organization-id` header. From those values, we can check if the user
is part of the organization. Once we have support for user API tokens,
we will update this middleware to accept a user API token (scoped to the
organization) to authenticate as well.
- New /account and GET /organizations endpoints to fetch the current
authenticated account and organizations associated with it.
- Since organization creation requires account-based authentication (no
organization context), a new POST /organizations endpoint was added to
allow authenticated accounts to create organizations.