refactor: split execution list into a separate endpoint (#244)
This pull request updates the state transitions for stage events. This
[whimsical](https://whimsical.com/event-states-2u4PCE3FLmjNYeybobKQTK)
of how the state of a stage event is updated while the event is being
processed.

### Main changes

- When a stage execution is started, the stage event no longer is moved
to `waiting(execution)` state. Instead, the stage event is moved to
`processed` state. The `waiting(X)` state is reserved to stage events
that are waiting on conditions to be met.
- The already existing `GET /stages/{id}/events` endpoint no longer
includes execution information.
- A new `GET /stages/{id}/executions` exposes information about stage
executions.
- A new `POST /stages/{id}/executions/{executionId}/cancel` was added to
allow stage executions to be cancelled.
- The `GET /event-sources` and `GET /stages` endpoints were updated to
include status information about the resources being returned. That
allows for us to implement the lazy loading without the need of the bulk
endpoints, which were removed.
- Cleaned up unused state reasons

---------

Co-authored-by: Pedro Foresti Leão <pforesti@renderedtext.com>
Co-authored-by: Lucas Pinheiro <lpinheiro@semaphore.io>