feat: track event rejections while processing events (#256)
### Issue

Currently, if an event is filtered, or something goes wrong while
applying the filters, or building the inputs, the event is rejected, but
we only log an error message, which doesn't allow users to see what
happened with their events that were rejected. That makes it hard to
troubleshoot issues with filters or inputs.

### Solution

This pull request addresses the issue in two ways:
- New state_reason and state_message fields in events table. Those
fields are used when processing event source filters. If something goes
wrong, or filters don't pass, the state_reason is updated accordingly
- New event_rejections table and GET /event_rejections endpoint to
expose them. These are used when processing filters for stages and
connection groups (and any future component type as well)

Note: the UI updates to display these event rejections for stages will
come in a later pull request