feat: manually emit event for stage (#306)
Allow the EmitEventModal component to be used for stage nodes as well.
Since the last emitted event by a stage is not exposed by the API, we
need to load it using the ListEvents endpoint.

### Other changes

- I was getting collation mismatch errors - `The database was created
using collation version 2.36, but the operating system provides version
2.41` - when running make `dev.setup` and `test.setup`. I suspect this
is due to updates in the Postgres image we are using in our docker
compose, so I updated the `db.create` Makefile target to always run
`REFRESH COLLATION VERSION` for `template1` and `postgres` databases
before creating new ones.
- Emit a RabbitMQ message when CreateEvent endpoint is used.

---------

Co-authored-by: Pedro Leão <60622592+forestileao@users.noreply.github.com>