fix: process executions in parallel in PendingExecutionsWorker (#339)
### Issue

The PendingExecutionsWorker processes executions sequentially. That
allows for a slow execution (HTTP call that takes a long time) to block
other executions to proceed.

### Solution

We combine a couple of things here:
- We use goroutines to process executions in parallel inside of the same
worker
- We use a semaphore from
[golang.org/x/sync/semaphore](https://pkg.go.dev/golang.org/x/sync/semaphore)
to control the number of concurrent goroutines running at the same time
- We use a transaction for each execution, locking the record with `FOR
UPDATE SKIP LOCKED` to avoid other workers from processing it at the
same time.