fix: use synced enforcer, transactional adapter, and LoadFilteredPolicy (#1058)
In order to use LoadFilteredPolicy, we need to rely on SyncedEnforcer.
We wouldn't be able to do that since we are using the transactional
adapter, but gorm adapter supports transaction, so we don't need to use
the transactional enforcer as well. Here, we switch to the
SyncedEnforcer, and rely solely on the adapter's support for
transactions. That allows us to use LoadFilteredPolicy instead of
LoadPolicy.