Auto-save works only on position change (#1344)
Fix for the bug introduced when adding auto-save for changing the node
position.
If I deleted component and then moved another one - auto save would kick
in.
This solves that - we want the user to have to click on `Save` for
destructive actions like deleting nodes and edges and adding new
connections