feat: github application (#1346)
An implementation of the Application interface for GitHub. The
implementation takes care of creating a GitHub app from manifest and
installing the newly created GitHub app into the required
account/organization. It also reacts to updates to the installation
(when repository is added or deleted, installation is suspended /
unsuspended, ...)

Initially, the GitHub application provides no components, and two
triggers:
- github.onPush - listens to push events
- github.onPullRequest - listens to pull requests

### Other changes

A new `any-predicate-list` field type was added to support having a list
of possible values for the refs to listen for in the `github.onPush`
trigger.