fix: simplify application interface webhook methods (#1456)
The logic of searching for an existing web hook that matches the
requested configuration is the same for all Application implementations.
The only thing that is different is the logic for comparing web hook
configurations. Due to that, we simplify the webhook-related methods by
doing the following:
- The RequestWebhook() methods lives only on the AppInstallationContext
- not on the Application interface - and it's how application
components/triggers implementations request a webhook from the the app
installation.
- No more ListWebhooks(), CreateWebhook() and AssociateWebhook() methods
in AppInstallationContext. That could be confusing to the user
implementing the component/trigger. Now, they only see a
RequestWebhook().
- A new CompareWebhookConfig() was added to the Application interface,
and it's how application implementations determine when to reuse web
hooks or not.

Signed-off-by: Lucas Pinheiro <lucas@superplane.com>